[FIX] little issues

+++ packages/chrome.sh
@@ -8,6 +8,11 @@ if (( $# > 0 )) && [ "$1" = 'description' ];then
     exit 0
 fi
 
+if [ -f "$(which google-chrome)" ];then
+    echo 'Already Installed'
+    exit 0
+fi
+
 sudo apt-get install -y wget
 
 cd /tmp
